Made inventory tests
Moved tests for the inventory to its own file away from test_avatar.py
